WA-268 PairingProcess component refactoring
